requirements.txt
asgiref==3.9.1
Django==5.2.5
django-allauth==65.11.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
